pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fic-expenses"
version = "0.2.0"
description = "TUI tool to manage expenses in Fatture in Cloud"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fattureincloud-python-sdk>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "textual>=0.50.0",
]

[project.scripts]
fic-expenses = "fic_expenses.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fic_expenses"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]
